AgroIpa-BackEnd/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

SQLAlchemy==2.0.25
asyncpg==0.29.0

alembic==1.13.1
python-dotenv==1.0.1

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

pydantic-settings==2.2.1
PyJWT==2.10.1